refactor: @AuthenticatedUser 어노테이션을 global 패키지로 이동
* 인증 관련 어노테이션을 global 계층으로 이동
* 재사용성과 의존성 분리를 강화
