+++ src/main/java/jeonseguard/backend/auth/infrastructure/resolver/AuthenticatedUserArgumentResolver.java
@@ -1,7 +1,7 @@
 package jeonseguard.backend.auth.infrastructure.resolver;
 
 import jakarta.servlet.http.HttpServletRequest;
-import jeonseguard.backend.auth.domain.annotation.AuthenticatedUser;
+import jeonseguard.backend.global.annotation.AuthenticatedUser;
 import jeonseguard.backend.auth.domain.repository.LogoutTokenStore;
 import jeonseguard.backend.auth.infrastructure.provider.JwtTokenProvider;
 import jeonseguard.backend.global.exception.error.*;
