+++ src/main/java/jeonseguard/backend/global/annotation/AuthenticatedUser.java
@@ -1,4 +1,4 @@
-package jeonseguard.backend.auth.domain.annotation;
+package jeonseguard.backend.global.annotation;
 
 import java.lang.annotation.*;
